requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.10.0
pandas>=2.0.0
fake-useragent>=1.4.0

# Telegram Bot dependencies
yandex-cloud-ml-sdk>=0.2.0
asyncio
aiohttp>=3.8.0

# Additional utilities
python-dotenv>=1.0.0